+++ packages/react-components/src/api/index.ts
@@ -53,7 +53,12 @@ import {
   updateScenario,
 } from './Scenarios/ScenariosApi';
 import { fetchSpreadsheetUsedRange, updateSpreadsheet } from './Spreadsheets/SpreadsheetsApi';
-import { fetchTimeseriesByGroup, fetchTimeseriesFullNames, fetchTimeseriesValues, fetchTimeseriesIdValues } from './Timeseries/TimeseriesApi';
+import {
+  fetchTimeseriesByGroup,
+  fetchTimeseriesFullNames,
+  fetchTimeseriesValues,
+  fetchTimeseriesIdValues,
+} from './Timeseries/TimeseriesApi';
 import {
   createUserGroup,
   deleteUserGroup,
@@ -61,6 +66,7 @@ import {
   updateUserGroups,
   updateUserGroupsForUser,
 } from './UserGroups/UserGroupsApi';
+import { fetchScalars } from './Scalars/ScalarsApi';
 
 export {
   fetchJsonDocuments,
@@ -118,4 +124,5 @@ export {
   fetchMapAnimationFiles,
   fetchFeatureCollection,
   fetchFeatureCollectionValues,
+  fetchScalars,
 };